LayoutTests/media/modern-media-controls/elapsed-time-support/elapsed-time-support-expected.txt

Testing the ElapsedTimeSupport behavior.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


timeupdate event was dispatched
PASS mediaController.controls.timeControl.elapsedTimeLabel.value is media.currentTime

PASS successfullyParsed is true

TEST COMPLETE
